Update snippet.db2formit.php
I was trying to get "notFoundRedirect" to work. It was not working at all for me. So i looked up the code and saw that the "else"-Option was stick to "if (is_array($where)) {". Shouldn't that belong to: "if ($dataobject = $modx->getObject($classname, $where)) {".
Now it is working fine. Am I wrong?